style and comments
