+++ src/main/scala/com/tpg/pnode/rabbit/RabbitConn.scala
@@ -8,8 +8,14 @@ import com.rabbitmq.client.ConnectionFactory
 import scala.concurrent.duration._
 import scala.language.postfixOps
 
+/** functionality to connect to RabbitMQ */
 object RabbitConn {
 
+  /**
+   * sets up an actor to handle a connection to a rabbit queue
+   * @param system the system the actor will live in
+   * @param handler a messages handler
+   */
   def setUpRabbit(system: ActorSystem, handler: ActorRef) = {
     val connFactory = new ConnectionFactory()
     connFactory.setUri("amqp://guest:guest@localhost/%2F")
