--- src/main/scala/com/tpg/pnode/rabbit/RabbitQueueSource.scala
@@ -1,39 +1,0 @@
-package com.tpg.pnode.rabbit
-
-import akka.actor.Actor
-import akka.stream.actor.{ActorPublisher, ActorPublisherMessage}
-
-case class RabbitMsg(val m: String)
-
-case object GetQueueSize
-
-class RabbitQueueSource extends Actor with ActorPublisher[RabbitMsg] {
-  import ActorPublisherMessage._
-
-  var rabbitMsgQueue:List[RabbitMsg] = List.empty
-
-  def receive = {
-
-    case GetQueueSize => sender() ! rabbitMsgQueue.size
-
-    case Request(demand) =>
-      if (demand > rabbitMsgQueue.size){
-        rabbitMsgQueue foreach (onNext)
-        rabbitMsgQueue = List.empty
-      }
-      else {
-        val (send, keep) = rabbitMsgQueue.splitAt(demand.toInt)
-        rabbitMsgQueue = keep
-        send foreach (onNext)
-      }
-
-
-    case rm:RabbitMsg =>
-      if (totalDemand == 0)
-        rabbitMsgQueue = rabbitMsgQueue :+ rm
-      else
-        onNext(rm)
-  }
-
-
-}
