+++ src/test/scala/com/tpg/pnode/RabbitQueueSourceActorTest.scala
@@ -6,11 +6,11 @@ import akka.stream.actor.ActorPublisher
 import akka.stream.actor.ActorPublisherMessage.Request
 import akka.stream.scaladsl.{Sink, Source}
 import akka.testkit._
-import com.tpg.pnode.rabbit.{GetQueueSize, RabbitMsg, RabbitQueueSource}
+import com.tpg.pnode.rabbit.RabbitQueueSource.{GetQueueSize, RabbitMsg, RabbitQueueSourceActor}
 import org.scalatest.{BeforeAndAfterAll, Matchers, WordSpecLike}
 
 
-class RabbitQueueSourceTest extends TestKit(ActorSystem("RabbitQueueSourceSpec"))
+class RabbitQueueSourceActorTest extends TestKit(ActorSystem("RabbitQueueSourceSpec"))
                             with DefaultTimeout with ImplicitSender
                             with WordSpecLike with Matchers with BeforeAndAfterAll {
 
@@ -20,10 +20,10 @@ class RabbitQueueSourceTest extends TestKit(ActorSystem("RabbitQueueSourceSpec")
   }
 
 
-  "An RabbitQueueSource" must {
+  "A RabbitQueueSourceActor" must {
 
     "start with an empty queue" in {
-      val ar = TestActorRef(new RabbitQueueSource)
+      val ar = TestActorRef(new RabbitQueueSourceActor)
       val a = ar.underlyingActor
 
       assert(a.rabbitMsgQueue.size == 0)
@@ -31,7 +31,7 @@ class RabbitQueueSourceTest extends TestKit(ActorSystem("RabbitQueueSourceSpec")
 
 
     "have one message queued after receiving a message" in {
-      val ar = TestActorRef(new RabbitQueueSource)
+      val ar = TestActorRef(new RabbitQueueSourceActor)
       val a = ar.underlyingActor
 
       ar ! new RabbitMsg("test")
@@ -43,7 +43,7 @@ class RabbitQueueSourceTest extends TestKit(ActorSystem("RabbitQueueSourceSpec")
     "empty its queue if requested by stream" in {
       implicit val materializer = ActorMaterializer()
 
-      val ra = system.actorOf(Props(classOf[RabbitQueueSource]))
+      val ra = system.actorOf(Props(classOf[RabbitQueueSourceActor]))
       val pub = ActorPublisher[RabbitMsg](ra)
       Source(pub).runWith(Sink.ignore)
